# Shuffle drops column "_partitions" in-place