--- a/dask/dataframe/core.py
+++ b/dask/dataframe/core.py
@@ -3676,11 +3676,11 @@ class DataFrame(_Frame):
         return {None: 0, "index": 0, "columns": 1}.get(axis, axis)
 
     @derived_from(pd.DataFrame)
-    def drop(self, labels=None, axis=0, columns=None, errors="raise"):
+    def drop(self, labels=None, axis=0, columns=None, errors="raise", inplace=False):
         axis = self._validate_axis(axis)
         if (axis == 1) or (columns is not None):
             return self.map_partitions(
-                M.drop, labels=labels, axis=axis, columns=columns, errors=errors
+                M.drop, labels=labels, axis=axis, columns=columns, errors=errors, inplace=inplace
             )
         raise NotImplementedError(
             "Drop currently only works for axis=1 or when columns is not None"