--- a/dask/dataframe/shuffle.py
+++ b/dask/dataframe/shuffle.py
@@ -248,7 +248,7 @@ def shuffle(df, index, shuffle=None, npartitions=None, max_branch=32, compute=No
         shuffle=shuffle,
         compute=compute,
     )
-    del df3["_partitions"]
+    df3.drop(columns="_partitions", inplace=True)
     return df3
 
 
@@ -270,7 +270,7 @@ def rearrange_by_divisions(df, column, divisions, max_branch=None, shuffle=None)
         npartitions=len(divisions) - 1,
         shuffle=shuffle,
     )
-    del df3["_partitions"]
+    df3.drop(columns="_partitions", inplace=True)
     return df3