Pandas' Dataframe support dropping a column in-place, which shuffle now uses.
Notice, cuDF will also support in-place drop: https://github.com/rapidsai/cudf/pull/3497

This PR **doubles** the performance of a distributed cuDF merge on a DGX1: 

### Before this PR (full copy)
```
Merge benchmark
--------------------------
Chunk-size  | 100000000
Frac-match  | 0.3
Ignore-size | 1.05 MB
Protocol    | ucx
Device(s)   | 1,2
==========================
Total time  | 2.51 s
==========================
(w1,w2)     | 25% 50% 75% (total nbytes)
--------------------------
(01,02)     | 6.77 GB/s 10.25 GB/s 10.36 GB/s (5.20 GB)
(02,01)     | 6.65 GB/s 10.06 GB/s 10.59 GB/s (5.20 GB)
```


### After this PR (drop in-place)
```
Merge benchmark
--------------------------
Chunk-size  | 100000000
Frac-match  | 0.3
Ignore-size | 1.05 MB
Protocol    | ucx
Device(s)   | 1,2
==========================
Total time  | 1.26 s
==========================
(w1,w2)     | 25% 50% 75% (total nbytes)
--------------------------
(01,02)     | 9.79 GB/s 9.96 GB/s 10.91 GB/s (5.60 GB)
(02,01)     | 9.74 GB/s 9.88 GB/s 10.69 GB/s (5.60 GB)
```

cc @mrocklin