@mrocklin 2019-12-02
Wow, it's great to see the 2x performance increase.

We should definitely find a way to do this in-place drop call, however, I don't think that we can support it reliably within the `dd.DataFrame.drop` method.  Tasks in Dask may run twice or may run in different orders than we expect, and so we generally don't modify inputs of a task in-place. 
 We can still call `my_cudf_dataframe.drop(..., inplace=True)`, but we'll have to call it immediately within some already existing task.  

Alternatively, is there a way within cudf to make `drop` faster?  Ideally calling `drop` would only be a change to metadata.  cc @kkraus14