@kkraus14 2019-12-02
> Alternatively, is there a way within cudf to make drop faster? Ideally calling drop would only be a change to metadata. cc @kkraus14

Both Pandas and cudf have drop return by deep copy by default which is expensive, but we want to match the semantics.

Currently the shuffle function isn't calling drop here, it's calling `del df['column']`. I would assume this runs inplace similar to `drop(..., inplace=True)`.