@kkraus14 2019-12-02
https://github.com/dask/dask/blob/master/dask/dataframe/core.py#L3296 This is the expensive line, could this use the `inplace` argument?