@mrocklin 2019-12-02
In this particular situation yes, but in full generality, unfortunately not.