@mrocklin 2019-12-02
Another option here would be to do this as an optimization pass.  If we ever find a blockwise operation that has a drop in it where that operation is not affecting the input then we add the `inplace=True` keyword.  This would affect all safe uses of drop, not just the ones here.

The downside of this is that high level optimizations are somewhat new, and so there aren't many examples on how to do them.  The best example of this today is probably the [`optimize_read_parquet_getitem`](https://github.com/dask/dask/blob/3ba4f6e43ecda75ed4ad19f42f3eba6c78a42d49/dask/dataframe/optimize.py#L36) optimization, which modifies read_parquet calls to select out only certain columns after the user has selected them, like this:

```python
df = dd.read_parquet(filename)
df = df[["x", "y"]]
```

```python
# becomes
df = dd.read_parquet(filename, columns=["x", "y"])
```