@mrocklin 2019-12-02
@rjzamora any interest in investigating dask dataframe optimizations?  

You would want to read through 

-  https://docs.dask.org/en/latest/high-level-graphs.html
-  The read_parquet optimization pointed to above
-  Maybe also https://github.com/dask/dask/pull/5451/files (although the read_parquet optimization is more relevant)

Then you would want to construct an optimization which looked through all Blockwise layers, found any that had an `M.drop` call and, if that call wasn't affecting inputs, added an `inplace=True` keyword to it.  

This would force you to learn about the high level graph structure a bit more, and also into the guts of subgraph callables.