@rjzamora 2019-12-02
> @rjzamora any interest in investigating dask dataframe optimizations?

Sounds good - Moving the `inplace=True` change into an optimization pass sounds like a good idea to me.  I will look into it.