@VibhuJawa 2019-12-02
FWIW, I feel the optimization pass will be super helpful across a bunch of more work-flows. 

In a lot of workflows to work around this issue I use column selection to work around it but a general solution will go a long way in simplifying things for us. (Please suggest if a better way is there)

```python
n_rows = 50_000_000
n_cols = 20
```
##### With drop
```python
df = dask_cudf.from_cudf(cudf.DataFrame({str(i):np.ones(shape=n_rows) for i in range(0,n_cols)}),npartitions=2)

st = time.time()
df = df.drop(columns=['0'])
x = len(df)
et = time.time()
print(f" time = {et - st}")
del df
```

```
time = 0.19231152534484863
```


##### With column selection

```python
df = dask_cudf.from_cudf(cudf.DataFrame({str(i):np.ones(shape=n_rows) for i in range(0,n_cols)}),npartitions=2)

st = time.time()
df = df[[str(i)for i in range(1,n_cols)]]
x = len(df)
et = time.time()
print(f" time = {et - st}")
del df
```
```
time = 0.0229644775390625
```

CC: @beckernick .