@mrocklin 2019-12-02
That's a really motivating example @VibhuJawa !  Thanks.

Yes, hopefully this helps other workloads.  Additionally, I hope that after doing this in one case we build up enough familiarity that we can do it in other cases as they arise as well.