@rjzamora 2019-12-02
@mrocklin - When parsing the layers of the high-level graph, the drop operations seem to have the generic `"subgraph_callable"` name, rather than something with "drop" in it.  Do you have any advice on the correct way to compare the underlying operation with the call we are looking for?  I assume we don't want to be using the key here.