@mrocklin 2019-12-02
Blockwise layers fuse everything into a single subgraph callable.  You will need to dive within this subgraph callable object and look at its task graph (called `.dsk` I think).  Within this graph you're going to want to search for tasks that call `M.drop`, look at their dependencies to verify that it's safe to insert the `inplace=True` keyword, and then do so, returning a new Blockwise layer with the modified subgraph callable with the modified `.dsk` attribute.  

I'd be more than happy to walk you through this over screenshare if you're interested.