@madsbk 2019-12-03
Created an issue to track this issue: https://github.com/dask/dask/issues/5664